+++ iterset.go
@@ -80,19 +80,6 @@ func (m MapSet[K, V]) Equal(keys iter.Seq[K]) bool {
 	return superset && len(m) == len(s)
 }
 
-// IsSubset returns whether every map key is present in keys.
-//
-// Related:
-//   - [MapSet.IsSuperset] if the keys were a map
-//   - [IsSubset] if the receiver was not a map
-//
-// Performance:
-//   - time: O(k)
-//   - space: O(min(m, k))
-func (m MapSet[K, V]) IsSubset(keys iter.Seq[K]) bool {
-	return len(m) == len(m.intersect(keys))
-}
-
 // IsSuperset returns whether all keys are present.
 //
 // Performance: