+++ legacy.go
@@ -1,0 +1,20 @@
+//go:build !go1.27
+
+package iterset
+
+import (
+	"iter"
+)
+
+// IsSubset returns whether every map key is present in keys.
+//
+// Related:
+//   - [MapSet.IsSuperset] if the keys were a map
+//   - [IsSubset] if the receiver was not a map
+//
+// Performance:
+//   - time: O(k)
+//   - space: O(min(m, k))
+func (m MapSet[K, V]) IsSubset(keys iter.Seq[K]) bool {
+	return len(m) == len(m.intersect(keys))
+}