+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `Sorted*` functions and intersection support multisets
 * Iteration optimized for slices
 
+### Added
+* Generic methods in Go >=1.27
+
 ## [0.5.0](https://github.com/coady/iterset/releases/tag/v0.5.0) - 2025-09-19
 ### Added
 * `Group`